requirements.txt
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
distro==1.8.0
exceptiongroup==1.2.0
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
openai==1.3.6
outcome==1.3.0.post0
packaging==23.2
playwright==1.40.0
pydantic==2.5.2
pydantic_core==2.14.5
pyee==11.0.1
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.31.0
selenium==4.15.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
tqdm==4.66.1
trio==0.23.1
trio-websocket==0.11.1
typing_extensions==4.8.0
urllib3==2.1.0
webdriver-manager==4.0.1
wsproto==1.2.0
